Help: Organize and revise 3.25 release notes
Add section headers similar to the 3.24 release notes and move each
individual bullet into an appropriate section.  Revise a few bullets.

Drop the `UseSWIG` bullet, as it is a simple bug fix.  Drop the
`LINK_LIBRARY` bullet, as it just polishes an existing feature,
and has a corresponding `versionadded` mark anyway.