Tests: Provide RunCMake.CommandLine compare_files data more robustly
The `lf` and `crlf` files need to have exact LF and CRLF newlines,
respectively.  Previously we used the Git `eol` attribute for this,
but occasionally the Git work tree ends up without the conversion.
It is unclear why that happens, so try to avoid the problem by
treating their content as binary instead.